+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - TBD
+
+### Changed
+
+- Landing page (/) and collections endpoint (/collections) now return a 500 if
+  create_indices has not been run or cannot connect to database.
+
 ## [3.3.0] - 2023-12-04
 
 ### Added
@@ -412,7 +419,7 @@ Initial release, forked from [sat-api](https://github.com/sat-utils/sat-api/tree
 
 Compliant with STAC 0.9.0
 
-<!-- [Unreleased]: https://github.com/stac-utils/stac-api/compare/v2.4.0...main -->
+[Unreleased]: https://github.com/stac-utils/stac-api/compare/v3.3.0...main
 [3.3.0]: https://github.com/stac-utils/stac-api/compare/v3.2.0...v3.3.0
 [3.2.0]: https://github.com/stac-utils/stac-api/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/stac-utils/stac-api/compare/v3.0.0...v3.1.0