+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - TBD
+## [3.4.0] - 2023-12-15
 
 ### Changed
 
@@ -421,7 +421,8 @@ Initial release, forked from [sat-api](https://github.com/sat-utils/sat-api/tree
 
 Compliant with STAC 0.9.0
 
-[Unreleased]: https://github.com/stac-utils/stac-api/compare/v3.3.0...main
+<!-- [Unreleased]: https://github.com/stac-utils/stac-api/compare/v3.3.0...main -->
+[3.4.0]: https://github.com/stac-utils/stac-api/compare/v3.3.0...v3.4.0
 [3.3.0]: https://github.com/stac-utils/stac-api/compare/v3.2.0...v3.3.0
 [3.2.0]: https://github.com/stac-utils/stac-api/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/stac-utils/stac-api/compare/v3.0.0...v3.1.0